use an ordered dict to store links inside visualizer
Robot is now just an alias for an ordered dict from a key type to Link,
and Link is just an alias for a vector of GeometryData. The key in the
ordered dict serves as the link name, which makes issues with duplicate
names less likely (and more obvious when they occur). Also, it is now
possible to pass draw() a dict from keys to transforms, which frees us
from having to remember the link order and also allows partial draws by
only sending transforms for some links.